ADD: 스키마 변경(codesquad-2023-group04#58)
-

+++ BE/src/main/resources/data/schma.sql
@@ -6,7 +6,7 @@ USE issue;
 
 CREATE TABLE `milestone`
 (
-    `milestone_id` INT          NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    `milestone_id` BIGINT       NOT NULL PRIMARY KEY AUTO_INCREMENT,
     `title`        VARCHAR(45)  NOT NULL,
     `description`  VARCHAR(300) NULL,
     `create_time`  DATETIME     NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -35,7 +35,7 @@ CREATE TABLE `issue`
     `update_time`  DATETIME    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     `is_open`      BIT         NOT NULL DEFAULT 1,
     `is_delete`    BIT         NOT NULL DEFAULT 0,
-    `milestone_id` INT         NULL,
+    `milestone_id` BIGINT      NULL,
     `author`       INT         NOT NULL,
 
     CONSTRAINT `fk_Issue_Milestone1`
@@ -91,9 +91,9 @@ CREATE TABLE `comment`
 
 CREATE TABLE `label_list`
 (
-    `label_list_id`  INT         NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `label_id` INT NOT NULL,
-    `issue_id` INT NOT NULL,
+    `label_list_id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `label_id`      INT NOT NULL,
+    `issue_id`      INT NOT NULL,
 
     CONSTRAINT `fk_Label_has_Issue_Label1`
         FOREIGN KEY (`label_id`)
@@ -111,9 +111,9 @@ CREATE TABLE `label_list`
 
 CREATE TABLE `assignee`
 (
-    `assignee_id`  INT         NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    `member_id` INT NOT NULL,
-    `issue_id`  INT NOT NULL,
+    `assignee_id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `member_id`   INT NOT NULL,
+    `issue_id`    INT NOT NULL,
 
     CONSTRAINT `fk_User_has_Issue_User1`
         FOREIGN KEY (`member_id`)
